Langchain-Cache/requirements.txt
python-dotenv==1.0.0
openai
cachetools
langchain
tiktoken